abstract result

+++ src/common.c
@@ -4,7 +4,7 @@
 int min(Vector* v)
 {
     int res = INT_MAX;
-    for (int i = 0; i < vector_length(v); i++) {
+    for (size_t i = 0; i < vector_length(v); i++) {
         int* value = vector_get(v, i);
         if (*value < res) {
             res = *value;
